+++ examples/basic-create-react-app/src/components/projects/graphs/Graph.js
@@ -2,6 +2,18 @@ import React from 'react';
 import _ from 'lodash';
 
 export default function Graph({graph}) {
+    const protocol = (name, protocol) => {
+        return <li>
+            {name}
+            <ul>
+                {
+                    _.keys(protocol).map((key) => (
+                        <li key={key}>{key}: {(protocol[key]).toString()}</li>
+                    ))
+                }
+            </ul>
+        </li>
+    };
     return (<ul>
         <li>
             {graph.name}
@@ -12,25 +24,28 @@ export default function Graph({graph}) {
                 <li>
                     <b>Connection:</b>
                     <ul>
-                        <li><b>Info:</b> {graph.connection.type}</li>
+                        <li><b>Type:</b> {graph.connection.type}</li>
+                        <li><b>Database type:</b> {graph.connection.databaseType}</li>
+                        <li><b>Database status:</b> {graph.connection.databaseStatus}</li>
                         <li>
                             <b>Info:</b>
                             <ul>
-                                {
-                                    _.keys(graph.connection.info).map((key) => (
-                                        <li key={key}>{key}: {(graph.connection.info[key] || "").toString()}</li>
-                                    ))
-                                }
+                                <li><b>Version:</b> {graph.connection.info.version}</li>
+                                <li><b>Edition:</b> {graph.connection.info.edition}</li>
                             </ul>
                         </li>
                         <li>
                             <b>Configuration</b>
                             <ul>
-                                {
-                                    _.keys(graph.connection.configuration).map((key) => (
-                                        <li key={key}>{key}: {(graph.connection.configuration[key] || "").toString()}</li>
-                                    ))
-                                }
+                                <li><b>Path:</b> {graph.connection.configuration.path}</li>
+                                <li>
+                                    <b>Protocols:</b>
+                                    <ul>
+                                        {protocol('bolt', graph.connection.configuration.protocols.bolt)}
+                                        {protocol('http', graph.connection.configuration.protocols.http)}
+                                        {protocol('https', graph.connection.configuration.protocols.https)}
+                                    </ul>
+                                </li>
                             </ul>
                         </li>
                     </ul>